requirements-dev.txt
# numpy <2 wheels stop at Python 3.12 (NumPy 2.1 adds Py 3.13 support)
numpy>=2.1,<3 ; python_version >= "3.13"
numpy<2       ; python_version <  "3.13"
duckdb>=0.9.2
feast>=0.38.0
schedule>=1.2.0              # lightweight job runner
python-dateutil>=2.8.2
pandas-ta==0.3.14            # realised-vol helper
tqdm>=4.66.0

finbert
schedule>=1.2
transformers==4.52.4
